+++ tests/test_adb_device.py
@@ -7,7 +7,8 @@
 #
 
 # pylint: disable=no-member
-"""Unit tests for adb commands"""
+""" Unit tests for adb commands. """
+
 import unittest
 import os
 import subprocess
@@ -20,32 +21,30 @@ TEST_DEVICE_ID = utils.get_test_device_id()
 
 class AdbDeviceTest(  # pylint: disable=too-many-public-methods
         unittest.TestCase):
-    """Adb device unit tests"""
+    """ Adb device unit tests. """
 
     @classmethod
     def setUpClass(cls):
-        """Download resources at the beginning of the tests"""
+        """ Start adb server and download tests resources at the beginning of
+        the tests. """
+        simpleadb.AdbServer()
         utils.download_resources()
 
     def setUp(self):
-        """Start adb server in each test"""
-        self.__adb = simpleadb.AdbServer()
+        """ Start adb server in each test. """
         utils.android_wait_for_emulator()
 
-    def tearDown(self):
-        """Kill adb server in each test"""
-        self.__adb.kill()
-
     def test_adb_devices_exists(self):
-        """Check if adb devices exists"""
-        devices = self.__adb.devices()
+        """Check if adb devices exists. """
+        adb_server = simpleadb.AdbServer()
+        devices = adb_server.devices()
         if not devices:
             self.fail('No adb devices found')
         test_device = simpleadb.AdbDevice(TEST_DEVICE_ID)
         self.assertTrue(test_device in devices)
 
     def test_device_eq_to_device_string(self):
-        """Test equal operator for adb device object"""
+        """Test equal operator for adb device object. """
         dev1 = simpleadb.AdbDevice('1234')
         dev2 = simpleadb.AdbDevice('1234')
         dev3 = simpleadb.AdbDevice('42')
@@ -53,7 +52,7 @@ class AdbDeviceTest(  # pylint: disable=too-many-public-methods
         self.assertNotEqual(dev1, dev3)
 
     def test_adb_device_is_str(self):
-        """Test qual operator for adb device and object and string devie id"""
+        """Test qual operator for adb device and string devie id. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
         self.assertEqual(TEST_DEVICE_ID, str(device))
 
@@ -66,7 +65,7 @@ class AdbDeviceTest(  # pylint: disable=too-many-public-methods
         self.assertTrue(device.is_available())
 
     def test_custom_adb_path_not_exists(self):
-        """Test custom adb binary path not exists"""
+        """Test custom adb binary path not exists. """
         device = simpleadb.AdbDevice(
             TEST_DEVICE_ID,
             path='dummy/path'
@@ -75,127 +74,123 @@ class AdbDeviceTest(  # pylint: disable=too-many-public-methods
             device.get_serialno()
 
     @pytest.mark.skipif(
-        utils.is_github_workflows_env(),
+        utils.is_github_workflows_env() or not utils.enable_root_tests(),
         reason='is_root() is "experimental" feature, may fail on emulator')
     def test_is_root_true_when_device_rooted(self):
-        """Check if device is rooted after adb root command"""
+        """Check if device is rooted after adb root command. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
         try:
             device.root(timeout_sec=10)
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)
         self.assertTrue(device.is_root())
 
+    @pytest.mark.skipif(
+        not utils.enable_root_tests(),
+        reason='Failing on not rootable devices')
     def test_adb_root_not_failing(self):
-        """Check adb root command not failing"""
+        """Check adb root command not failing. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
         try:
             device.root()
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)
 
     def test_get_id_returns_correct_id(self):
-        """Check if get_id is equal to test device id"""
+        """Check if get_id is equal to test device id. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
         test_device_id = device.get_id()
         self.assertEqual(test_device_id, TEST_DEVICE_ID)
 
     def test_get_serialno_returns_correct_id(self):
-        """Check if get_serialno is equal to test device id"""
+        """Check if get_serialno is equal to test device id. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
-        if not device.is_root():
-            device.root()
         test_device_id = device.get_serialno()
         self.assertEqual(test_device_id, TEST_DEVICE_ID)
 
     def test_input_tap_not_failing(self):
-        """Check if input tap is not failing"""
+        """Check if input tap is not failing. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
-        if not device.is_root():
-            device.root()
         try:
             device.tap(1, 1)
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)
 
     def test_input_swipe_not_failing(self):
-        """Check if input swipe is not failing"""
+        """Check if input swipe is not failing. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
-        if not device.is_root():
-            device.root()
         try:
             device.swipe(1, 1, 2, 2)
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)
 
     def test_screenshot_if_screenshot_file_exists(self):
-        """Check if screenshot file exists after screencap"""
+        """Check if screenshot file exists after screencap. """
         filepath = './screenshot.png'
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
         try:
             device.screencap(local=filepath)
             device.screencap()
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)
         self.assertTrue(os.path.isfile(filepath))
 
     def test_install_apk_if_success(self):
         """Check if install and uninstall apk are not failing"""
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
-        if not device.is_root():
-            device.root()
-        device.remount()
         try:
             device.install(utils.DUMMY_APK_NAME)
             device.uninstall(utils.DUMMY_PACKAGE_NAME)
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)
 
+    @pytest.mark.skipif(
+        not utils.enable_root_tests(),
+        reason='Failing on not rootable devices')
     def test_set_setprop_is_not_failing(self):
         """Check if setprop is not failing"""
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
-        if not device.is_root():
-            device.root()
         try:
             device.setprop("dummy_prop", "true")
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)
 
+    @pytest.mark.skipif(
+        not utils.enable_root_tests(),
+        reason='Failing on not rootable devices')
     def test_getprop(self):
-        """Verify if property value is correct using getprop command"""
+        """Verify if property value is correct using getprop command. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
         prop_name = 'dummy_prop'
         prop_val = 'true'
-        if not device.is_root():
-            device.root()
         try:
             device.setprop(prop_name, prop_val)
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)
         self.assertEqual(prop_val, device.getprop(prop_name))
 
     @pytest.mark.skipif(
         utils.is_github_workflows_env(),
+        not utils.enable_root_tests(),
         reason='Failing on emulator')
     def test_verity(self):
-        """Check if verity command is not failing"""
+        """Check if verity command is not failing. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
         device.root()
         device.remount()
         try:
             device.enable_verity(True)
             device.enable_verity(False)
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            print("ERR: ", err)
+            self.fail(err)
 
     def test_push_pull(self):
-        """Verify if file exists after push/pull command"""
+        """Verify if file exists after push/pull command. """
         filename = 'dummy_file'
         dest = '/sdcard/'
 
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
-        if not device.is_root():
-            device.root()
         os.system('touch ' + filename)
         try:
             device.push(filename, dest)
@@ -203,11 +198,11 @@ class AdbDeviceTest(  # pylint: disable=too-many-public-methods
             self.assertFalse(os.path.isfile(filename))
             device.pull(dest + filename)
             self.assertTrue(os.path.isfile(filename))
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)
 
     def test_remove(self):
-        """Check if file was removed after rm command"""
+        """Check if file was removed after rm command. """
         filename = 'test_remove_dummy_file'
         dest = '/sdcard/'
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
@@ -216,14 +211,14 @@ class AdbDeviceTest(  # pylint: disable=too-many-public-methods
             device.push(filename, dest)
             os.remove(filename)
             device.rm(dest + filename)
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)
 
         with self.assertRaises(simpleadb.AdbCommandError):
             device.pull(dest + filename)
 
     def test_remove_not_existing_file_failure(self):
-        """Check if rm command failed if file not exists"""
+        """Check if rm command failed if file not exists. """
         filename = '/sdcard/no_existing_file'
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
 
@@ -231,93 +226,89 @@ class AdbDeviceTest(  # pylint: disable=too-many-public-methods
             device.rm(filename)
 
     def test_get_state_returns_correct_state(self):
-        """Check adb get state command"""
+        """Check adb get state command. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
-        if not device.is_root():
-            device.root()
         state = device.get_state()
         self.assertEqual(state, 'device')
 
     def test_dump_logcat(self):
-        """Test dump logcat"""
+        """Test dump logcat. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
         log = device.dump_logcat()
         self.assertNotEqual(log, None)
 
     def test_dump_logcat_from_buffer(self):
-        """Test dump logcat"""
+        """Test dump logcat. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
         log = device.dump_logcat('main')
         self.assertNotEqual(log, None)
 
     def test_clear_logcat(self):
-        """Test clear logcat"""
+        """Test clear logcat. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
         try:
             device.clear_logcat()
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)
 
     def test_clear_logcat_buf(self):
-        """Test clear logcat from main buffer"""
+        """Test clear logcat from main buffer. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
         try:
             device.clear_logcat('main')
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)
 
     @pytest.mark.skipif(
         utils.is_github_workflows_env(),
         reason='Failing on emulator')
     def test_device_is_available(self):
-        """Test if device is available"""
+        """Test if device is available. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
-        if not device.is_root():
-            device.root()
         self.assertTrue(device.is_available())
 
     def test_no_available(self):
-        """Test if device not available"""
+        """Test if device not available. """
         device = simpleadb.AdbDevice('dummy_id')
         self.assertFalse(device.is_available())
 
     def test_wait_for_device(self):
-        """Test wait for device"""
+        """Test wait for device. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
         try:
             device.wait_for_device()
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)
 
     def test_wait_for_device_failed(self):
-        """Wait for device failed after timeout if device not exists"""
+        """Wait for device failed after timeout if device not exists. """
         with self.assertRaises(subprocess.TimeoutExpired):
             device = simpleadb.AdbDevice('dummy-device')
             device.wait_for_device(1)
 
     def test_wait_for_device_timeout(self):
-        """Test wait for device with custom timeout"""
+        """Test wait for device with custom timeout. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
-        if not device.is_root():
-            device.root()
         try:
             device.wait_for_device(5)
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)
 
     def test_adb_shell(self):
-        """Test adb shell input command"""
+        """Test adb shell input command. """
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
-        device.root()
         try:
             device.shell('input text 42')
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)
 
+    @pytest.mark.skipif(
+        not utils.enable_root_tests(),
+        reason='Failing on not rootable devices')
     def test_unroot(self):
         """Test unroot"""
         device = simpleadb.AdbDevice(TEST_DEVICE_ID)
         try:
             device.unroot()
-        except simpleadb.AdbCommandError:
-            self.fail("AdbCommandError unexpectedly")
+        except simpleadb.AdbCommandError as err:
+            self.fail(err)